+++ src/main/java/org/tinyradius/packet/RadiusResponse.java
@@ -12,7 +12,7 @@ public class RadiusResponse extends BaseRadiusPacket {
     /**
      * Builds a Radius packet with the given type, identifier and attributes.
      * <p>
-     * Use {@link RadiusPackets#createResponse(Dictionary, int, int, byte[], List)}
+     * Use {@link RadiusPackets#createResponse(Dictionary, byte, byte, byte[], List)}
      * where possible as that automatically creates Access/Accounting
      * variants as required.
      *
@@ -22,7 +22,7 @@ public class RadiusResponse extends BaseRadiusPacket {
      * @param authenticator can be null if creating manually
      * @param attributes    list of RadiusAttribute objects
      */
-    public RadiusResponse(Dictionary dictionary, int type, int identifier, byte[] authenticator, List<RadiusAttribute> attributes) {
+    public RadiusResponse(Dictionary dictionary, byte type, byte identifier, byte[] authenticator, List<RadiusAttribute> attributes) {
         super(dictionary, type, identifier, authenticator, attributes);
     }
