+++ src/main/java/org/tinyradius/packet/util/PacketType.java
@@ -7,36 +7,36 @@ import java.util.HashMap;
  */
 public class PacketType {
 
-    public static final int ACCESS_REQUEST = 1; // RFC 2865
-    public static final int ACCESS_ACCEPT = 2;
-    public static final int ACCESS_REJECT = 3;
-    public static final int ACCESS_CHALLENGE = 11;
+    public static final byte ACCESS_REQUEST = 1; // RFC 2865
+    public static final byte ACCESS_ACCEPT = 2;
+    public static final byte ACCESS_REJECT = 3;
+    public static final byte ACCESS_CHALLENGE = 11;
 
-    public static final int ACCOUNTING_REQUEST = 4; // RFC 2866
-    public static final int ACCOUNTING_RESPONSE = 5;
-    public static final int ACCOUNTING_STATUS = 6; // aka Interim Accounting
-    public static final int ACCOUNTING_MESSAGE = 10;
+    public static final byte ACCOUNTING_REQUEST = 4; // RFC 2866
+    public static final byte ACCOUNTING_RESPONSE = 5;
+    public static final byte ACCOUNTING_STATUS = 6; // aka Interim Accounting
+    public static final byte ACCOUNTING_MESSAGE = 10;
 
-    public static final int PASSWORD_REQUEST = 7;
-    public static final int PASSWORD_ACCEPT = 8;
-    public static final int PASSWORD_REJECT = 9;
+    public static final byte PASSWORD_REQUEST = 7;
+    public static final byte PASSWORD_ACCEPT = 8;
+    public static final byte PASSWORD_REJECT = 9;
 
-    public static final int STATUS_SERVER = 12; // RFC 5997
-    public static final int STATUS_CLIENT = 13;
+    public static final byte STATUS_SERVER = 12; // RFC 5997
+    public static final byte STATUS_CLIENT = 13;
 
-    public static final int DISCONNECT_REQUEST = 40; // RFC 5176
-    public static final int DISCONNECT_ACK = 41;
-    public static final int DISCONNECT_NAK = 42;
-    public static final int COA_REQUEST = 43;
-    public static final int COA_ACK = 44;
-    public static final int COA_NAK = 45;
+    public static final byte DISCONNECT_REQUEST = 40; // RFC 5176
+    public static final byte DISCONNECT_ACK = 41;
+    public static final byte DISCONNECT_NAK = 42;
+    public static final byte COA_REQUEST = 43;
+    public static final byte COA_ACK = 44;
+    public static final byte COA_NAK = 45;
 
-    public static final int STATUS_REQUEST = 46;
-    public static final int STATUS_ACCEPT = 47;
-    public static final int STATUS_REJECT = 48;
-    public static final int RESERVED = 255;
+    public static final byte STATUS_REQUEST = 46;
+    public static final byte STATUS_ACCEPT = 47;
+    public static final byte STATUS_REJECT = 48;
+    public static final byte RESERVED = (byte) 255;
 
-    private static final HashMap<Integer, String> typeNames = new HashMap<Integer, String>() {{
+    private static final HashMap<Byte, String> typeNames = new HashMap<Byte, String>() {{
         put(ACCESS_REQUEST, "Access-Request");
         put(ACCESS_ACCEPT, "Access-Accept");
         put(ACCESS_REJECT, "Access-Reject");
@@ -62,7 +62,7 @@ public class PacketType {
         put(RESERVED, "Reserved");
     }};
 
-    public static String getPacketTypeName(int code) {
+    public static String getPacketTypeName(byte code) {
         return typeNames.getOrDefault(code, "Unknown (" + code + ")");
     }
 }
